Session added patient

+++ src/main/java/com/project/clinic/clinic/controllers/DoctorController.java
@@ -2,14 +2,17 @@ package com.project.clinic.clinic.controllers;
 
 import com.project.clinic.clinic.daos.DoctorDao;
 import com.project.clinic.clinic.daos.DoctorScheduleDao;
+import com.project.clinic.clinic.models.Admin;
 import com.project.clinic.clinic.models.Doctor;
 import com.project.clinic.clinic.models.DocSchedule;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.print.Doc;
 import java.util.List;
 
 @Controller
@@ -22,9 +25,13 @@ public class DoctorController {
     DoctorDao dao;
 
     @GetMapping("/doctorcreate")
-    public ModelAndView createDoctorGet() {
-        return new ModelAndView("/doctor/doctorcreate", "doctor", new Doctor());
-    }
+    public ModelAndView createDoctorGet(HttpSession session) {
+            Admin admin=(Admin) session.getAttribute("admin");
+        if (admin==null){
+            return new ModelAndView("redirect:/login");
+        }
+            return new ModelAndView("/doctor/doctorcreate", "doctor", new Doctor());
+        }
 
     @PostMapping("/doctorcreate")
     public ModelAndView createDoctorPost(@ModelAttribute Doctor doctor) {
@@ -33,23 +40,39 @@ public class DoctorController {
     }
 
     @GetMapping("/doctorview")
-    public String doctorView(Model model) {
-        List<Doctor> doctors = dao.findAll();
-        model.addAttribute("doctors", doctors);
-        return "/doctor/doctorview";
-    }
+    public String doctorView(Model model,HttpSession session) {
+        Doctor doctor = (Doctor) session.getAttribute("doctor");
+        Admin admin = (Admin) session.getAttribute("admin");
+        if  ( admin == null && doctor == null){
+            return "redirect:/login";
+        }
+            List<Doctor> doctors = dao.findAll();
+            model.addAttribute("doctors", doctors);
+            return "/doctor/doctorview";
+        }
 
     @GetMapping("/delete/doctor/{doctor_id}")
-    public String deletedoctor(@PathVariable("doctor_id") Long doctor_id) {
+    public String deletedoctor(@PathVariable("doctor_id") Long doctor_id,HttpSession session) {
+        Admin admin=(Admin) session.getAttribute("admin");
+        if (admin ==null){
+            return "redirect:/login";
+        }
         dao.deleteById(doctor_id);
         return "redirect:/doctorview";
     }
 
     @GetMapping("/edit/doctor/{doctor_id}")
-    public ModelAndView doctoredit(@PathVariable("doctor_id") Long doctor_id) {
+    public ModelAndView doctoredit(@PathVariable("doctor_id") Long doctor_id,HttpSession session) {
+        Admin admin=(Admin) session.getAttribute("admin");
+        if (admin ==null){
+            return new ModelAndView("redirect:/login");
+        }else{
         Doctor doctor = dao.findById(doctor_id).orElseThrow();
         return new ModelAndView("/doctor/doctoredit", "doctorBean", doctor);
     }
+
+    }
+
 
     @PostMapping("/update/doctor")
     public String updateAdmin(@ModelAttribute("doctorBean") Doctor doctor) {
@@ -57,12 +80,12 @@ public class DoctorController {
         return "redirect:/doctorview";
     }
 
-
-
-
-
     @GetMapping("/doctorschedulecreate")
-    public ModelAndView createDoctorScheduleGet() {
+    public ModelAndView createDoctorScheduleGet(HttpSession session) {
+        Admin admin=(Admin) session.getAttribute("admin");
+        if (admin ==null){
+            return new ModelAndView("redirect:/login");
+        }
         return new ModelAndView("/doctorschedule/doctorschedulecreate", "schedule", new DocSchedule());
     }
 
@@ -73,20 +96,32 @@ public class DoctorController {
     }
 
     @GetMapping("/doctorscheduleview")
-    public String doctorScheduleView(Model model) {
+    public String doctorScheduleView(Model model,HttpSession session) {
+        Admin admin=(Admin) session.getAttribute("admin");
+        if (admin ==null){
+            return "redirect:/login";
+        }
         List<DocSchedule> docSchedules = doctordao.findAll();
         model.addAttribute("docSchedules", docSchedules);
         return "/doctorschedule/doctorscheduleview";
     }
 
     @GetMapping("/delete/doctorschedule/{schedule_id}")
-    public String deleteDoctorSchedule(@PathVariable("schedule_id") Long schedule_id) {
+    public String deleteDoctorSchedule(@PathVariable("schedule_id") Long schedule_id,HttpSession session) {
+        Admin admin=(Admin) session.getAttribute("admin");
+        if (admin ==null){
+            return "redirect:/login";
+        }
         doctordao.deleteById(schedule_id);
         return "redirect:/doctorscheduleview";
     }
 
     @GetMapping("/edit/doctorschedule/{schedule_id}")
-    public ModelAndView doctorScheduleEdit(@PathVariable("schedule_id") Long schedule_id) {
+    public ModelAndView doctorScheduleEdit(@PathVariable("schedule_id") Long schedule_id,HttpSession session) {
+        Admin admin=(Admin) session.getAttribute("admin");
+        if (admin ==null){
+            return new ModelAndView("redirect:/login");
+        }
         DocSchedule schedule = doctordao.findById(schedule_id).orElseThrow();
         return new ModelAndView("/doctorschedule/doctorscheduleedit", "scheduleBean", schedule);
     }
