+++ src/main/java/com/project/clinic/clinic/controllers/PatientController.java
@@ -3,6 +3,7 @@ package com.project.clinic.clinic.controllers;
 import com.project.clinic.clinic.daos.PatientDao;
 import com.project.clinic.clinic.models.Admin;
 import com.project.clinic.clinic.models.Patient;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -29,24 +30,36 @@ public class PatientController {
     }
 
     @GetMapping("/patientview")
-    public String patientview(Model model){
+    public String patientview(Model model,HttpSession session){
+        Admin admin=(Admin) session.getAttribute("admin");
+        if (admin ==null){
+            return "redirect:/login";
+        }
         List<Patient> patients = dao.findAll();
         model.addAttribute("patients",patients);
         return "/patient/patientview";
     }
 
     @GetMapping("/delete/patient/{patient_id}")
-    public String deletePatient(@PathVariable("patient_id")Long patient_id){
+    public String deletePatient(@PathVariable("patient_id")Long patient_id, HttpSession session){
+        Admin admin=(Admin) session.getAttribute("admin");
+        if (admin ==null){
+            return "redirect:/login";
+        }
         dao.deleteById(patient_id);
         return  "redirect:/patientview";
     }
 
-    @GetMapping("/edit/patient/{patient_id}")
-    public ModelAndView editPatient(@PathVariable("patient_id")Long patient_id){
+    @GetMapping("/patient/edit/{patient_id}")
+    public ModelAndView editPatient(@PathVariable("patient_id")Long patient_id,HttpSession session){
+        Patient checkPatient =(Patient) session.getAttribute("patient");
+        if (checkPatient == null){
+            return new ModelAndView("redirect:/login");
+        }
         Patient patient =dao.findById(patient_id).orElseThrow();
         return new ModelAndView("/patient/patientedit","patientBean",patient);
     }
-    @PostMapping("/update/patient")
+    @PostMapping("/patient/update")
     public String updatePatient(@ModelAttribute("patientBean")Patient patient){
         dao.save(patient);
         return "redirect:/patientview";
